Moved ID and MENU_ACTION_TYPE to companion object
